Move overflow menu items to navigtion drawer
