+++ app/src/main/java/io/ipoli/android/MainActivity.java
@@ -1,12 +1,16 @@
 package io.ipoli.android;
 
+import android.Manifest;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
 import android.support.annotation.StringRes;
 import android.support.design.widget.AppBarLayout;
 import android.support.design.widget.CoordinatorLayout;
 import android.support.design.widget.NavigationView;
 import android.support.design.widget.Snackbar;
+import android.support.v4.app.ActivityCompat;
 import android.support.v4.app.Fragment;
 import android.support.v4.content.ContextCompat;
 import android.support.v4.widget.DrawerLayout;
@@ -18,7 +22,10 @@ import android.view.View;
 import android.widget.ImageView;
 import android.widget.ProgressBar;
 import android.widget.TextView;
+import android.widget.Toast;
 
+import com.facebook.share.model.AppInviteContent;
+import com.facebook.share.widget.AppInviteDialog;
 import com.github.sundeepk.compactcalendarview.CompactCalendarView;
 import com.squareup.otto.Subscribe;
 
@@ -32,10 +39,14 @@ import javax.inject.Inject;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import io.ipoli.android.app.BaseActivity;
+import io.ipoli.android.app.events.CalendarPermissionResponseEvent;
+import io.ipoli.android.app.events.ContactUsTapEvent;
 import io.ipoli.android.app.events.CurrentDayChangedEvent;
 import io.ipoli.android.app.events.EventSource;
 import io.ipoli.android.app.events.FeedbackTapEvent;
+import io.ipoli.android.app.events.InviteFriendEvent;
 import io.ipoli.android.app.events.ScreenShownEvent;
+import io.ipoli.android.app.events.SyncCalendarRequestEvent;
 import io.ipoli.android.app.events.UndoCompletedQuestEvent;
 import io.ipoli.android.app.rate.RateDialog;
 import io.ipoli.android.app.rate.RateDialogConstants;
@@ -61,6 +72,8 @@ import io.ipoli.android.quest.fragments.OverviewFragment;
 import io.ipoli.android.quest.fragments.RepeatingQuestListFragment;
 import io.ipoli.android.quest.persistence.QuestPersistenceService;
 import io.ipoli.android.reward.fragments.RewardListFragment;
+import io.ipoli.android.tutorial.TutorialActivity;
+import io.ipoli.android.tutorial.events.ShowTutorialEvent;
 
 public class MainActivity extends BaseActivity implements View.OnClickListener, NavigationView.OnNavigationItemSelectedListener {
 
@@ -382,9 +395,24 @@ public class MainActivity extends BaseActivity implements View.OnClickListener, 
                 changeCurrentFragment(new RewardListFragment(), R.string.title_fragment_rewards);
                 break;
 
+            case R.id.invite_friends:
+                eventBus.post(new InviteFriendEvent());
+                inviteFriend();
+                break;
+            case R.id.sync_calendars:
+                checkForCalendarPermission();
+                break;
+            case R.id.tutorial:
+                eventBus.post(new ShowTutorialEvent());
+                startTutorial();
+                break;
             case R.id.feedback:
                 eventBus.post(new FeedbackTapEvent());
                 EmailUtils.send(this, getString(R.string.feedback_email_subject), getString(R.string.feedback_email_chooser_title));
+                break;
+            case R.id.contact_us:
+                eventBus.post(new ContactUsTapEvent());
+                EmailUtils.send(this, getString(R.string.contact_us_email_subject), getString(R.string.contact_us_email_chooser_title));
                 break;
         }
 
@@ -392,4 +420,50 @@ public class MainActivity extends BaseActivity implements View.OnClickListener, 
 
         return true;
     }
+
+    private void startTutorial() {
+        Intent intent = new Intent(this, TutorialActivity.class);
+        startActivity(intent);
+    }
+
+    private void inviteFriend() {
+        if (AppInviteDialog.canShow()) {
+            AppInviteContent content = new AppInviteContent.Builder()
+                    .setApplinkUrl(Constants.FACEBOOK_APP_LINK)
+                    .setPreviewImageUrl(Constants.FACEBOOK_INVITE_IMAGE_URL)
+                    .build();
+            AppInviteDialog.show(this, content);
+        } else {
+            Toast.makeText(this, R.string.show_invite_failed, Toast.LENGTH_LONG).show();
+        }
+    }
+    private void checkForCalendarPermission() {
+        if (ContextCompat.checkSelfPermission(this,
+                Manifest.permission.READ_CALENDAR)
+                != PackageManager.PERMISSION_GRANTED) {
+
+            ActivityCompat.requestPermissions(this,
+                    new String[]{Manifest.permission.READ_CALENDAR},
+                    Constants.READ_CALENDAR_PERMISSION_REQUEST_CODE);
+        } else {
+            eventBus.post(new SyncCalendarRequestEvent(EventSource.OPTIONS_MENU));
+            Toast.makeText(this, R.string.import_calendar_events_started, Toast.LENGTH_SHORT).show();
+        }
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        if (requestCode == Constants.READ_CALENDAR_PERMISSION_REQUEST_CODE) {
+            if (grantResults.length > 0
+                    && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                eventBus.post(new CalendarPermissionResponseEvent(CalendarPermissionResponseEvent.Response.GRANTED, EventSource.OPTIONS_MENU));
+                eventBus.post(new SyncCalendarRequestEvent(EventSource.TUTORIAL));
+                Toast.makeText(this, R.string.import_calendar_events_started, Toast.LENGTH_SHORT).show();
+            } else if(grantResults.length > 0
+                    && grantResults[0] == PackageManager.PERMISSION_DENIED) {
+                eventBus.post(new CalendarPermissionResponseEvent(CalendarPermissionResponseEvent.Response.DENIED, EventSource.OPTIONS_MENU));
+            }
+        }
+    }
+
 }
