+++ app/src/main/java/io/ipoli/android/app/BaseActivity.java
@@ -1,34 +1,15 @@
 package io.ipoli.android.app;
 
-import android.Manifest;
-import android.content.Intent;
-import android.content.pm.PackageManager;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
 import android.view.Menu;
-import android.view.MenuItem;
-import android.widget.Toast;
 
-import com.facebook.share.model.AppInviteContent;
-import com.facebook.share.widget.AppInviteDialog;
 import com.squareup.otto.Bus;
 import com.trello.rxlifecycle.components.support.RxAppCompatActivity;
 
 import javax.inject.Inject;
 
-import io.ipoli.android.Constants;
 import io.ipoli.android.R;
-import io.ipoli.android.app.events.CalendarPermissionResponseEvent;
-import io.ipoli.android.app.events.ContactUsTapEvent;
-import io.ipoli.android.app.events.EventSource;
-import io.ipoli.android.app.events.InviteFriendEvent;
-import io.ipoli.android.app.events.SyncCalendarRequestEvent;
-import io.ipoli.android.app.utils.EmailUtils;
-import io.ipoli.android.tutorial.TutorialActivity;
-import io.ipoli.android.tutorial.events.ShowTutorialEvent;
 
 /**
  * Created by Venelin Valkov <venelin@curiousily.com>
@@ -53,73 +34,5 @@ public class BaseActivity extends RxAppCompatActivity {
     public boolean onCreateOptionsMenu(Menu menu) {
         getMenuInflater().inflate(R.menu.main_menu, menu);
         return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.action_contact_us:
-                eventBus.post(new ContactUsTapEvent());
-                EmailUtils.send(this, getString(R.string.contact_us_email_subject), getString(R.string.contact_us_email_chooser_title));
-                break;
-            case R.id.action_show_tutorial:
-                eventBus.post(new ShowTutorialEvent());
-                startTutorial();
-                break;
-            case R.id.action_invite_friend:
-                eventBus.post(new InviteFriendEvent());
-                inviteFriend();
-                break;
-            case R.id.action_sync_android_calendar:
-                checkForCalendarPermission();
-                break;
-        }
-        return super.onOptionsItemSelected(item);
-    }
-
-    private void inviteFriend() {
-        if (AppInviteDialog.canShow()) {
-            AppInviteContent content = new AppInviteContent.Builder()
-                    .setApplinkUrl(Constants.FACEBOOK_APP_LINK)
-                    .setPreviewImageUrl(Constants.FACEBOOK_INVITE_IMAGE_URL)
-                    .build();
-            AppInviteDialog.show(this, content);
-        } else {
-            Toast.makeText(this, R.string.show_invite_failed, Toast.LENGTH_LONG).show();
-        }
-    }
-
-    protected void startTutorial() {
-        Intent intent = new Intent(this, TutorialActivity.class);
-        startActivity(intent);
-    }
-
-    private void checkForCalendarPermission() {
-        if (ContextCompat.checkSelfPermission(this,
-                Manifest.permission.READ_CALENDAR)
-                != PackageManager.PERMISSION_GRANTED) {
-
-            ActivityCompat.requestPermissions(this,
-                    new String[]{Manifest.permission.READ_CALENDAR},
-                    Constants.READ_CALENDAR_PERMISSION_REQUEST_CODE);
-        } else {
-            eventBus.post(new SyncCalendarRequestEvent(EventSource.OPTIONS_MENU));
-            Toast.makeText(this, R.string.import_calendar_events_started, Toast.LENGTH_SHORT).show();
-        }
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        if (requestCode == Constants.READ_CALENDAR_PERMISSION_REQUEST_CODE) {
-            if (grantResults.length > 0
-                    && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                eventBus.post(new CalendarPermissionResponseEvent(CalendarPermissionResponseEvent.Response.GRANTED, EventSource.OPTIONS_MENU));
-                eventBus.post(new SyncCalendarRequestEvent(EventSource.TUTORIAL));
-                Toast.makeText(this, R.string.import_calendar_events_started, Toast.LENGTH_SHORT).show();
-            } else if(grantResults.length > 0
-                    && grantResults[0] == PackageManager.PERMISSION_DENIED) {
-                eventBus.post(new CalendarPermissionResponseEvent(CalendarPermissionResponseEvent.Response.DENIED, EventSource.OPTIONS_MENU));
-            }
-        }
     }
 }
