+++ app/src/main/res/drawable/ic_tutorial_white_24dp.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+        android:width="24dp"
+        android:height="24dp"
+        android:viewportWidth="24.0"
+        android:viewportHeight="24.0">
+    <path
+        android:fillColor="#FFFFFFFF"
+        android:pathData="M7,24h2v-2L7,22v2zM11,24h2v-2h-2v2zM15,24h2v-2h-2v2zM16,0.01L8,0C6.9,0 6,0.9 6,2v16c0,1.1 0.9,2 2,2h8c1.1,0 2,-0.9 2,-2L18,2c0,-1.1 -0.9,-1.99 -2,-1.99zM16,16L8,16L8,4h8v12z"/>
+</vector>
