+++ app/src/main/res/menu/main_menu.xml
@@ -7,29 +7,4 @@
         android:menuCategory="secondary"
         android:title="@string/help_explain_screen"
         app:showAsAction="never" />
-
-    <item
-        android:id="@+id/action_invite_friend"
-        android:menuCategory="secondary"
-        android:title="@string/invite_friend"
-        app:showAsAction="never" />
-
-    <item
-        android:id="@+id/action_sync_android_calendar"
-        android:menuCategory="secondary"
-        android:title="@string/sync_android_calendars"
-        app:showAsAction="never" />
-
-    <item
-        android:id="@+id/action_show_tutorial"
-        android:menuCategory="secondary"
-        android:title="@string/show_tutorial"
-        app:showAsAction="never" />
-
-    <item
-        android:id="@+id/action_contact_us"
-        android:menuCategory="secondary"
-        android:title="@string/contact_us"
-        app:showAsAction="never" />
-
 </menu>
