+++ app/src/main/res/values/strings.xml
@@ -140,7 +140,7 @@
     <string name="show_invite_failed">Oops, please, try again later!</string>
     <string name="cannot_edit_completed_quests">Completed quests cannot be edited</string>
     <string name="overview_title">Your Journey Ahead</string>
-    <string name="sync_android_calendars">Import calendars</string>
+    <string name="sync_android_calendars">Sync calendars</string>
     <string name="import_calendar_events_started">Your events will be available soon…</string>
     <string name="widget_journey_name">Daily Journey</string>
     <string name="add">ADD</string>
